chore(action): use ms for parsing miliseconds
